% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{conformal_scores}
\alias{conformal_scores}
\title{Conformal prediction for any given scores}
\usage{
conformal_scores(ste, s, y, labels)
}
\arguments{
\item{ste}{the list of scores for test samples corresponding to each of the training class.}

\item{s}{the vector scores for training samples corresponding to its own class.}

\item{y}{labels for the training data.}

\item{labels}{labels for the K classes.}
}
\value{
prediction_tr: conformal scores for the training samples below to its own class.

prediction_tr: conformal scores for the test samples corresponding to each of the training class.
}
\description{
Conformal prediction for any given scores
}
