% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.conformal}
\alias{evaluate.conformal}
\title{Evaluate  the conformal prediction results}
\usage{
evaluate.conformal(prediction, yte, labels, alpha = 0.05)
}
\arguments{
\item{prediction}{the m by K prediction matrix containing the constructed conformal p-value for m test samples and K training classes.}

\item{yte}{class labels for the test samples.}

\item{labels}{the K class labels for the training samples.}

\item{alpha}{the targeted type I error.}
}
\value{
a result table with the columns being the classes in the test samples, and the rows being the classes in the training samples. 
The entry at row j and column k represents the percent of samples in class j assigned label k.
}
\description{
Evaluate  the conformal prediction results
}
\examples{
### Examples
\dontrun{
data(mnist); 
xtrain = mnist[['data']][['x']]; ytrain = mnist[['data']][['y']]; 
xtest = mnist[['data_te']][['x']]; ytest=mnist[['data_te']][['y']]; 
#########data splitting##############
set.seed(123)
foldid = sample(1:2, length(ytrain), replace = TRUE); foldid_te = sample(1:2,length(ytest), replace = TRUE)
xtrain1 = xtrain[foldid==1,]; xtrain2 = xtrain[foldid==2,]; 
ytrain1 = ytrain[foldid==1]; ytrain2=ytrain[foldid==2]
xtest1 = xtest[foldid_te ==1,]; xtest2 = xtest[foldid_te==2,]; 
labels = sort(unique(ytrain)) 
#########example using random forest##############
require(ranger)
bcops = BCOPS(ranger, xtrain1, ytrain1, xtest1, xtrain2, ytrain2, xtest2, labels, formula = TRUE, prediction_only = FALSE)
prediction.conformal = matrix(NA, ncol = length(labels), nrow = length(ytest))
prediction.conformal[foldid_te==1,] = bcops[[1]];
prediction.conformal[foldid_te==2,] = bcops[[2]];
evaluate.conformal(prediction.conformal, ytest, labels, 0.05)
}
}
