% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{a small subset of MNIST handwritten digit data}
\format{{list}
\describe{
  \item{data}{training data}
  \item{data_te}{test data}
}}
\usage{
data(mnist)
}
\description{
a small subset of MNIST handwritten digit data
}
\examples{
data(mnist)
}
\references{
Hinton, Geoffrey E., and Ruslan R. Salakhutdinov. "Reducing the dimensionality of data with neural networks." science 313, no. 5786 (2006): 504-507.
}
\keyword{datasets}
