% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.conformal}
\alias{evaluate.conformal}
\title{Evaluate  the conformal prediction results}
\usage{
evaluate.conformal(prediction, yte, labels, alpha = 0.05)
}
\arguments{
\item{prediction}{the n by K prediction matrix containing the constructed conformal p-value for n test samples and K training classes.}

\item{yte}{actual class labels for the test samples.}

\item{labels}{the K class labels for the training samples.}

\item{alpha}{the targeted type I error.}
}
\value{
a result table with the columns being the classes in the test samples, and the rows being the classes in the training samples. 
The entry at row j and column k represents the percent of samples in class j assigned lable k.
}
\description{
Evaluate  the conformal prediction results
}
