% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{Proximity analysis example with prostate cancer data}
\format{{list}
\describe{
  \item{data}{training data}
  \item{data_te}{test data}
}}
\usage{
data(mnist)
}
\description{
The data is stored as a list. The training data with name data_train and the test data
with name data_test. It consists of $n=67$ training observations and 30 test observations. There are eight predictors.
The response is the log PSA for men who had  prostate cancer surgery.
}
\examples{
data(mnist)
}
\references{
Hinton, Geoffrey E., and Ruslan R. Salakhutdinov. "Reducing the dimensionality of data with neural networks." science 313, no. 5786 (2006): 504-507.
}
\keyword{datasets}
